# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Auth",
        "version": "1.0.0"
    },
    "paths": {
        "/api/keys/create": {
            "post": {
                "tags": [
                    "API Keys"
                ],
                "summary": "Create new API key",
                "operationId": "7f1ca4c19dab3c8aed969a5b10ba6c73",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "permissions",
                                    "expiry"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "wallet-service"
                                    },
                                    "permissions": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "enum": [
                                                "deposit",
                                                "transfer",
                                                "read"
                                            ]
                                        }
                                    },
                                    "expiry": {
                                        "description": "1H, 1D, 1M, or 1Y",
                                        "type": "string",
                                        "example": "1D"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "API key created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "api_key": {
                                                    "description": "SAVE THIS! It's shown only once",
                                                    "type": "string"
                                                },
                                                "expires_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/keys/rollover": {
            "post": {
                "tags": [
                    "API Keys"
                ],
                "summary": "Rollover expired API key",
                "operationId": "a10fd5469582f87e8c5e37222e22cc94",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "expired_key_id",
                                    "expiry"
                                ],
                                "properties": {
                                    "expired_key_id": {
                                        "type": "string"
                                    },
                                    "expiry": {
                                        "type": "string",
                                        "example": "1M"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "API key rolled over"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/keys": {
            "get": {
                "tags": [
                    "API Keys"
                ],
                "summary": "List all API keys",
                "operationId": "bd82250cacaf0cb5e68f5fb5f41ef508",
                "responses": {
                    "200": {
                        "description": "List of API keys"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/keys/{id}": {
            "delete": {
                "tags": [
                    "API Keys"
                ],
                "summary": "Revoke API key",
                "operationId": "57329fb496fcb6b78d89590dd620112a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "API key revoked"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/wallet/transfer": {
            "post": {
                "tags": [
                    "Transfer"
                ],
                "summary": "Transfer to another wallet",
                "operationId": "10cc1674827d231cf759a964377d8e86",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "wallet_number",
                                    "amount"
                                ],
                                "properties": {
                                    "wallet_number": {
                                        "type": "string",
                                        "example": "45123456789012"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "example": 3000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transfer successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "reference": {
                                                    "type": "string"
                                                },
                                                "amount": {
                                                    "type": "number"
                                                },
                                                "recipient_wallet": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "apiKey": {
                "type": "apiKey",
                "description": "API Key Authentication",
                "name": "x-api-key",
                "in": "header"
            }
        }
    },
...
}